Move `Guardian` to `guard.rs`
